convert columns: usize -> u16

+++ src/pages/address.rs
@@ -106,7 +106,7 @@ fn create_bmp_base64_qr(message: &str) -> Result<String, Error> {
     Ok(to_data_url(cursor.into_inner(), "image/bmp"))
 }
 
-pub fn text_page(address: &Address, page: &str, col: usize) -> Result<String, Error> {
+pub fn text_page(address: &Address, page: &str, col: u16) -> Result<String, Error> {
     let mut s = convert_text_html_string(page, col);
     s.push_str("\n");
     s.push_str(&create_string_qr(&address.to_qr_uri())?);
